api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video_processor"
version = "0.1.0"
description = "Video processing pipeline for Echo project"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Echo Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies
    "requests>=2.20.0",
    "google-cloud-secret-manager>=2.16.0",
    "google-cloud-storage==2.10.0",
    "vertexai",
    "ffmpeg-python",
    "google-cloud-aiplatform",
    "gunicorn",
    
    # YouTube API dependencies
    "google-api-python-client>=2.80.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "python-dotenv>=1.0.0",
    
    # Cloud Functions dependencies
    "functions-framework",
    
    # Data validation
    "pydantic==1.10.14",
    
    # Authentication
    "supabase>=2.0.0",
    
    # FastAPI for new API implementation
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
]

[project.optional-dependencies]
dev = [
    # Testing dependencies
    "pytest==7.4.0",
    "pytest-mock==3.11.1",
    "pytest-cov==4.1.0",
    
    # Development tools
    "black==24.1.1",
    "ruff==0.2.1",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/parkerrex/echo"
"Bug Tracker" = "https://github.com/parkerrex/echo/issues"
    
[tool.setuptools]
packages = ["video_processor"]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=video_processor --cov-report=term-missing"
markers = [
    "integration: integration tests that rely on external services or emulators",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "W", "N", "B", "A", "C4", "SIM", "ERA"]
ignore = ["E203"]

[tool.ruff.isort]
known-first-party = ["video_processor"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
